requirements.txt
# requirements.txt - Dependencies for ESS Energy Arbitrage Simulation

# Core scientific computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Mathematical optimization
pulp>=2.6.0

# Data visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Data input/output
openpyxl>=3.0.7
xlrd>=2.0.1
requests>=2.25.0

# Configuration handling
pyyaml>=5.4.0

# Optional: Enhanced optimization solvers
# gurobipy>=9.5.0  # Commercial - requires license
# cplex>=20.1.0    # Commercial - requires license

# Development and testing (optional)
pytest>=6.2.0
jupyter>=1.0.0

# Additional utilities
tqdm>=4.62.0  # Progress bars
colorama>=0.4.4  # Colored output

